ClassRefChecker functionality complete and has debugging statements
I put the verify function in the identifier visit function to add classes
that do not exist yet. I put the notify function in the identifier function's
declaring ST_CLASS(_EXTEND) case to remove classes that now exist. I call
checkClassesExisted in Goal(). The rest of the changes are in preparation for
storing methods in ClassVarData instead of in scopes.