equals and hashCode functions for all SymbolData classes, method arg typechecking done
MessageSend has a lot of code for a lot of conditions. It has to check that every type
in its expression list is correct and if the method exists. More checks for backpatching
to come, but assuming the method does exist and has been put in the symboltable already,
the method passes successfully.