vapor code for accessing field variables works now
moved getFieldVar code from symbolTable to scope.java. Tested that index code was outputting correct
indices given the locations of field variables. Outputted Vapor code. Should be index of field * 4 + 4
for the method table pointer, which it is. Setting synthTempVar to the temp that contains the field.
Seems good now.